+++ NEWS.md
@@ -34,6 +34,8 @@
 
 ## Features and bugfixes
 
+* External backtraces in error chains are now separately displayed (#1098).
+
 * Trace capture now better handles wrappers of calling handler in case
   of rethrown chained errors.